FaceClassify/requirements.txt
pytorch
numpy
torchvision